pyproject.toml
[project]
name = "netra-sdk"
version = "0.1.0"
description = "A Python SDK for AI application observability that provides OpenTelemetry-based monitoring, tracing, and PII protection for LLM and vector database applications. Enables easy instrumentation, session tracking, and privacy-focused data collection for AI systems in production environments."
authors = [
    {name = "Sooraj Thomas",email = "sooraj@keyvalue.systems"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "opentelemetry-api>=1.34.0,<2.0.0",
    "opentelemetry-sdk>=1.34.0,<2.0.0",
    "traceloop-sdk>=0.40.7,<0.41.0",
    "presidio-analyzer>=2.2.358,<3.0.0",
    "presidio-anonymizer>=2.2.358,<3.0.0",
    "flair>=0.15.1,<0.16.0",
]

[[tool.poetry.packages]]
include = "netra"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
autopep8 = "^2.3.2"
pytest = "^8.4.0"
commitizen = "^4.8.2"
pre-commit = "^4.2.0"
autoflake = "^2.3.1"
isort = "^6.0.1"
black = "^25.1.0"
mypy = "^1.16.0"

# ---------------------------
# Formatting: black
# ---------------------------
[tool.black]
line-length = 120
target-version = ['py310']
skip-string-normalization = false

# ---------------------------
# Import Sorting: isort
# ---------------------------
[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

# ---------------------------
# Type Checking: mypy
# ---------------------------
[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = true
check_untyped_defs = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
